+++ recast-java/src/main/java/io/improbable/ste/recast/PolyPointResult.java
@@ -10,7 +10,7 @@ public class PolyPointResult extends Structure {
     public static class ByReference extends PolyPointResult implements Structure.ByReference {}
     // NOTE: These two should match Types.kt
     public int status;
-    public int polyRef;
+    public long polyRef;
     public float[] point = new float[3];
 
     @Override
