+++ recast-java/src/main/kotlin/io/improbable/ste/recast/Types.kt
@@ -5,5 +5,5 @@ import com.sun.jna.Pointer
 typealias RcCompactHeightfield = Pointer
 typealias RcPolyMeshDetail = Pointer
 typealias DtStatus = Int
-typealias DtPolyRef = Int
+typealias DtPolyRef = Long
 typealias DtQueryFilter = Pointer
